requirements.txt
selenium
sh
ipgetter
psutil
hjsystemtools
html2text
requests
beautifulsoup4
requests[socks]